gene record overview - todo chart
